continued flexbox work
I continued working with flexbox and modifying the css code to make it function properly. I have the boxes working as intended, still working to implement the varaiable items in the actual event boxes. current task, getting an image of the date on the left side of the event's box and a description on the right side of the image.